+++ package.json
@@ -42,7 +42,8 @@
         "react-dom": "^17.0.2",
         "react-error-boundary": "^3.1.4",
         "react-helmet": "^6.1.0",
-        "react-leaflet": "2.6.3",
+        "react-leaflet": "^3.2.5",
+        "react-leaflet-draw": "^0.19.8",
         "react-query": "^3.34.2",
         "react-router-dom": "^5.2.0",
         "react-router-last-location": "^2.0.1",
@@ -56,7 +57,7 @@
         "url-search-params-polyfill": "^8.1.1"
     },
     "scripts": {
-        "build-tailwind": "npx tailwindcss-cli@latest build -i src/css/tailwind.src.css -c  tailwind.config.js -o src/css/tailwind.css",
+        "build-tailwind": "npx tailwindcss-cli@latest build -i src/css/tailwind.src.css -c tailwind.config.js -o src/css/tailwind.css",
         "prestart": "npm run-script build-tailwind",
         "prebuild": "npm run-script build-tailwind",
         "start": "craco start",
@@ -92,7 +93,9 @@
         "@testing-library/react": "^12.1.2",
         "@types/d3": "^7.1.0",
         "@types/leaflet": "^1.7.8",
+        "@types/leaflet-draw": "^1.0.5",
         "@types/lodash.clonedeep": "^4.5.6",
+        "@types/lodash.isequal": "^4.5.5",
         "@types/proj4leaflet": "^1.0.6",
         "@types/quill": "^2.0.9",
         "@types/react-helmet": "^6.1.5",
