--- src/components/Leaflet/LeafletDrawController/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LeafletDrawController'
