+++ src/components/Leaflet/LeafletSearchInput/LeafletSearchInput.tsx
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { ChangeEvent, RefObject, useRef, useState } from 'react'
+import { ChangeEvent, forwardRef, useRef, useState } from 'react'
 import { toast } from 'react-toastify'
 
 import { getLookupData, getSuggestData } from '@/api/axiosLocatieserver'
@@ -10,14 +10,13 @@ import debounce from '@/utils/debounce'
  */
 
 interface LeafletSearchInputProps {
-    reference: RefObject<any>
     mapPanTo: (lng: number, lat: number, type: string) => void
 }
 
-const LeafletSearchInput = ({
-    reference,
-    mapPanTo,
-}: LeafletSearchInputProps) => {
+const LeafletSearchInput = forwardRef<
+    HTMLInputElement,
+    LeafletSearchInputProps
+>(({ mapPanTo }, ref) => {
     const [searchQuery, setSearchQuery] = useState('')
     const [queryData, setQueryData] = useState([])
 
@@ -96,7 +95,7 @@ const LeafletSearchInput = ({
                 break
             case 'previous':
                 if (parseInt(currentIndex) === 1) {
-                    reference.current.select()
+                    typeof ref !== 'function' && ref?.current?.select()
                     return
                 }
                 newIndex = parseInt(currentIndex) - 1
@@ -116,7 +115,7 @@ const LeafletSearchInput = ({
             <input
                 className="w-64 h-10 px-5 py-3 ml-3 text-sm leading-tight text-gray-700 border-none rounded appearance-none focus:outline-none"
                 type="text"
-                ref={reference}
+                ref={ref}
                 placeholder="Zoeken op de kaart"
                 onChange={handleChange}
                 value={searchQuery}
@@ -172,6 +171,6 @@ const LeafletSearchInput = ({
             ) : null}
         </>
     )
-}
+})
 
 export default LeafletSearchInput
