--- src/components/Leaflet/LeafletViewer/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LeafletViewer'
