+++ src/components/Leaflet/index.ts
@@ -1,15 +1,11 @@
-import LeafletController from './LeafletController'
-import LeafletDrawController from './LeafletDrawController'
+import LeafletMap from './LeafletMap'
 import LeafletRevisionOverview from './LeafletRevisionOverview'
 import LeafletSearchInput from './LeafletSearchInput'
 import LeafletTinyViewer from './LeafletTinyViewer'
-import LeafletViewer from './LeafletViewer'
 
 export {
-    LeafletController,
-    LeafletDrawController,
     LeafletRevisionOverview,
     LeafletSearchInput,
     LeafletTinyViewer,
-    LeafletViewer,
+    LeafletMap,
 }
