+++ src/css/styles.scss
@@ -1145,12 +1145,12 @@ mark {
     }
 }
 
-#leaflet-tiny-viewer {
-    .leaflet-control {
-        margin-right: 0px;
-    }
-    .leaflet-control {
-        margin-top: 0px;
+.leaflet-map {
+    .leaflet-right {
+        .leaflet-control {
+            margin-right: 0px;
+            margin-top: 0px;
+        }
     }
     .leaflet-control-zoom.leaflet-control {
         margin-right: 10px;
